rd-agent: Simplify linux tarball check on startup
This used to lead to spurious download attempts and long stalls when rd-agent
gets restarted soon after it started - SIGTERM kills the tar verification
command triggering the download path which doesn't exit until SIGKILL gets
delivered on timeout.

There's no need to verify the tarball on each startup. Make sure we rename only
after the file is ready and that it isn't empty.