hashd: normalize cpu consumption against memory access size
cpu_ratio used to scale on top of file accesses which can lead to very
different behaviors when file and anon access balance is changed. Let's
apply it against total memory access instead of file only. cpu_ratio of 1.0
now means that the same number of bytes as total memory access is
checksummed.