agent, demo: sync command issues
agent didn't use to sync with agent after it updates the cmd.json file. This
means that there can be multiple command updates before the agent has the
chance to notice the changes. This meant that e.g. when on/off/on sequence
gets requested quickly, the agent might miss the off. We've been working
around this with using different ids for sys/sideloads in the same page but
we now also want to reset hashd in some cases.

This patch implements seq number based sync mechanism and make the demo ui
wait for agent to catch up before each "on" command.