bench: replace rctx.first_report() with rctx.report_sample()
The first report after agent starts up isn't the most relevant to the
testing. Instead sample report at ~30s before the end of the benchmark. This
should capture the report while benchmark is in progress and should e.g.
capture that iocost is disabled while iocost benchmark is in progress.