hashd, bench: use preset hash size and chunk pages instead of benchmarking them
hashd used cpu-single benchmark to determine file_size_mean and chunk_pages.
This makes the workload vary in terms of CPU and IO usage depending on CPU
performance making it difficult to compare results across different
machines. Instead, let's default to using the default values for these two
parameters so that the only things which are benchmarked by hashd are max
RPS and memory footprint.

* hashd now accepts --bench-cpu-single to enable the old cpu-single bench
  for determining hash and chunk sizes. This might still be useful in CPUs
  which are extremely slow or fast.

* hashd now accepts --bench-chunk-pages to specify the chunk pages
  explicitly.

* Hash and chunk sizes are now selected from one of the bench result, the
  default value or the explicitly specified value.

* resctl-bench is updated accordingly. The storage bench now accepts
  chunk-pages parameter and uses default values for hash size and log bps.