hashd: apply Boyu's hashd parameters
Boyu Ni <boyuni@fb.com> tuned hashd parameters on a 64G machine w/ Toshiba
XG5 SSD so that it's cpu, memory and IO behaviors match those of part 6 web
task. While still an approximation, this makes rd-hashd behave in a way
which is a lot easier to interpret in light of actual workloads. Let's use
Boyu's parameters as the default.

With the new parameters, we tend to get higher RPS on the same
configuration. Let's double the fake-cpu-load target RPS for storage
benchmark to ensure that storage benchmark can generate enough IOs even for
really fast SSDs.