bench: iocost-tune: improve line fitting
The assumption was that there will be single inflection point at the higher
end and the code fitted the data points with the combination of a slope and
a top flat line. However, for example, a device's latency may stay flat on
the lower end too when the device becomes under-saturated and showing its
inherent minimum latencies.

Improve line fitting by trying four different strategies and pick the one
with the smallest error - simple lineal regression without inflection point,
two segment line with a single inflection point at the top or bottom, and
three segment line with two inflection points.