bench: iocost-qos: study the whole run period
iocost-qos retries a few times after a storage or protection sub-bench
failure. It used to only include the periods from successful runs in vrate
and IO latency studies. We want to account for what happened for the failed
runs. Update the code so that it includes the whole run period while
studying the results.

While at it, rename main_period[s] to period[s].