bench: don't trim result file at the end
The staged updates are already maintaining perfectly fine result file. The
write at the end is there to trim the result file so that it only contains
results that's specified for the current run. This isn't necessary and makes
it easy to accidentally lose bench results, especially the ones which were
created by nested benches.

Remove the trimming overwrite at the end.