bench: separate out study into its own phase
Currently, Job::run() is supposed to run the actual bench and then analyze
it and produce the final result. This turned out to be a bit too restrictive
for developing and improving benchmarks which take really long to finish.
It'd be useful to be able to iterate on the analysis part without needing to
re-run the whole benchmark.

Let's separate out study into its own phase. run() is now expected to
produce a record which only contains enough information to run the actual
analysis. study() takes the record and produces the final result which is
then fed to format().

The benchmarks aren't yet updated and treat the record as the final result
for now.