bench: iocost-tune: implement no signal backstop for [a]MOF=max solutions
When [a]MOF is a flat line, the max solutions used to pick the min vrate. On
these devices, throttling usually doesn't buy much and the only effect of
sticking to min could be losing raw bandwidth. Let's use lat-imp left
inflection point as the backstop for these cases - ie. if the [a]MOF line
doesn't provide any signal, the solution is the maximum vrate which yields
the minimum lat-imp. lat-imp is a good summary of IO latencies, so we're
saying that we'll only go lower if we can observe latency gains.