bench: scale rps_max w/ mem_share for hashd fake cpu load hashd benchs
Fake cpu load hashd benchmarks are used to evaluate storage performance in
isolation. Let's scale the max RPS along with mem_share so that larger
memory profile means tougher IO workload so that users have the ability to
scale up the benchmark if necessary.

The scaling factor is set up so that the max RPS remains at 2000 when the
default 16G memory profile.