Server-side automation for results submission
This change allows resctl-bench to be executed as an AWS lambda
function to validate result files and create issues on github
for them.

AWS lambda expects a `bootstrap' binary and does not give it any
arguments. For that reason we first look at our executable name to
decide whether we should be running as a lambda function and re-exec
ourselves with the appropriate arguments.

The lambda function's goal is to validate and submit benchmark results
to github as issues. The benchmark file is uploaded to S3.

Signed-off-by: Gustavo Noronha Silva <gustavo.noronha@collabora.com>