bench: add upload subcommand
The upload subcommand is the client for the AWS lambda function. It
validates the result file and uploads it to the function for an issue to
be created automatically.

Optional parameters allow the user to include their email address or
github handle in the issue.

The lambda has been improved to use that data and to be better at
identifying duplicates.

Signed-off-by: Gustavo Noronha Silva <gustavo.noronha@collabora.com>