bench: iocost-tune: limit the number of candidate points during line fitting
The previous commit made the number of candidate points dependent on the
number of data points, which usually isn't too high. However, to avoid
pathological case when there are a lot of data points, limit the number of
candidate points by skipping over points which are too close. The limit is
currently set at 0.25% which should limit the number of candidate points
under 400.