bench: iocost-tune: implement DataShape::SinglePeak and apply it to AMOF
On some devices, amof peaks in the middle which the existing line fittings
couldn't describe reasonably. Implement new SinglePeak line fitting which
supports more complicated shapes and apply it to AMOF. It's likely that we
can replace other line fittings with this one while keeping the shape
constraints. Let's not rock that boat right for now.